requirements.txt
Flask
pandas
openpyxl
gunicorn
python-dotenv
authlib